♻️ [Refactoring]: ツール実装をZodベースのバリデーションに移行
- 各ツールでZodスキーマを使用するように更新
- validateメソッドを削除し、バリデーション責任を分離
- JsonSchema.from()を使用してinputSchemaを自動生成
- 新しいArgsファイルからスキーマとタイプをインポート
